quantizeNoiseShaping(): finilize the algorithm
this patch completes the algoritm that the block diagram
in 83ec28d0 describes. ATM it does not have frequency
adjustment on the quantization noise. while this is
recommended in theory, in practice we don't really
want to adjust the spectrum of the error with a static
transfer function but rather randomize it.

the algorithm is the result of a good amount of trial and error.
it's almost twice as efficient as the already included FS error
diffusion based dithering.

the PRNG used is a tiny, fast "xorshift" which is led to output
values in the 0.0..1.0 range. the noise levels are bits-per-channel
dependent, and can be adjusted via the 'noiseLevel' vector.

combining the current channel sample with the quantization noise
from the previous one produces diffusion or smearing of the higher
spectrum frequencies, due to the feedback.
this process however does also produce static distortion,
which can be observed as patterns. at that point, combining it the
PRNG samples further more diffuses the error.

practicly a combination between the FS alogrithm and this one
is possible, but i don't think this will end up being too optimal.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>