add sh/cmd variants of a 'buildrelease' script
buildrelease's purpose is to generate zip files of the current
release and the used AIR SDK.

it will generate files in the lines of:
./release/ta-gen_v1.5_air17.0.zip

7za is required on both osx (sh) and win32 (cmd).

on win32 a callback hack is used, because we want to be able
to call immediatelly the buildrelease.cmd body only after
build.cmd is executed. not a very nice solution.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>